Вынос компонентов в отдельные папки и файлы. - по favorites
